add Controller Service

+++ src/main/java/com/telran/controller/ProductController.java
@@ -11,7 +11,7 @@ import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
-@RequestMapping("/garden")
+@RequestMapping("/products")
 public class ProductController {
 
     private final ProductService service;
@@ -28,8 +28,8 @@ public class ProductController {
 
     @GetMapping("/all")
     public List<Product> getAll() {
-
-        return service.getAll();
+        List<Product> all = service.getAll();
+        return new ResponseEntity<>(all, HttpStatus.OK).getBody();
     }
 
     @GetMapping("/searchByName")
@@ -55,15 +55,18 @@ public class ProductController {
     public ResponseEntity<Product> updateProduct(@RequestBody Product product) {
         boolean isUpdated = service.updateProduct(product);
         if (isUpdated) {
-            return new ResponseEntity<>(product, HttpStatus.OK);
+            return new ResponseEntity<>(isUpdated ? HttpStatus.OK : HttpStatus.NOT_FOUND);
         } else {
-            return new ResponseEntity<>(product, HttpStatus.NO_CONTENT);
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
 
     @DeleteMapping
     public ResponseEntity<Product> deleteProduct(@RequestBody Product product) {
+        if (product.getName() == null || product.getName().isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         service.remove(product);
-        return new ResponseEntity<>(product, HttpStatus.ACCEPTED);
+        return new ResponseEntity<>(product, HttpStatus.OK);
     }
 }
